Small change to prevent matrix continuously disconnecting
Powers on the LED matrix as soon as a connection is established.  This will
function correctly with the python buildhat library as instead of attempting to
turn off the matrix when .off() is called, it simply clears the matrix display.

If off is called via minicom the matrix will then not function.